requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
httpx==0.28.1
openai==1.93.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydub==0.25.1


